Image::Colormap introduced.
git-svn-id: svn+ssh://cv077/home/ueshiba/svnroot/trunk/src@1998 3ef1b8b6-fb57-41bf-88d9-9a5bd220bf38

+++ Testprog/simd/transform/main.cc
@@ -61,6 +61,6 @@ doJob()
 int
 main()
 {
-    TU::simd::doJob<double, int8_t, int8_t, int16_t, int32_t>();
+    TU::simd::doJob<float, int8_t, int8_t, int16_t, int32_t>();
     return 0;
 }
